Move SetFDLimit to tests
