+++ testonly/hammer/hammer.go
@@ -31,7 +31,6 @@ import (
 	"github.com/google/trillian/monitoring"
 	"github.com/google/trillian/testonly"
 	"github.com/google/trillian/types"
-	"golang.org/x/sys/unix"
 )
 
 const (
@@ -166,26 +165,8 @@ func (c MapConfig) String() string {
 		c.MapID, c.EPBias, c.Operations, c.EmitInterval, c.RetryErrors)
 }
 
-// SetFDULimit sets the soft limit on the maximum number of open file descriptors.
-// See http://man7.org/linux/man-pages/man2/setrlimit.2.html
-func SetFDLimit(uLimit uint64) error {
-	var rLimit unix.Rlimit
-	err := unix.Getrlimit(unix.RLIMIT_NOFILE, &rLimit)
-	if err != nil {
-		return err
-	}
-	if uLimit > rLimit.Max {
-		return fmt.Errorf("Could not set FD limit to %v. Must be less than the hard limit %v", uLimit, rLimit.Max)
-	}
-	rLimit.Cur = uLimit
-	return unix.Setrlimit(unix.RLIMIT_NOFILE, &rLimit)
-}
-
 // HitMap performs load/stress operations according to given config.
 func HitMap(ctx context.Context, cfg MapConfig) error {
-	if err := SetFDLimit(2048); err != nil {
-		return err
-	}
 	var firstErr error
 
 	if cfg.MapID == 0 {
